requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.0.1
cffi==1.17.1
cfgv==3.4.0
click==8.1.8
cryptography==44.0.0
decorator==4.4.2
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
factory_boy==3.3.1
Faker==28.4.1
fastapi==0.114.2
fastapi-cli==0.0.5
fastapi-pagination==0.12.31
fastapi-utils==0.7.0
filelock==3.16.1
h11==0.14.0
httptools==0.6.4
identify==2.6.5
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
jwt==1.3.1
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
passlib==1.7.4
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==3.5.0
psutil==5.9.8
psycopg2==2.9.9
pycparser==2.22
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
Pygments==2.18.0
PyJWT==2.10.1
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqladmin==0.19.0
SQLAlchemy==2.0.34
sqlalchemy_database==0.1.2
sqlmodel==0.0.22
starlette==0.38.6
traitlets==5.14.3
typer==0.15.1
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
virtualenv==20.26.6
watchfiles==0.24.0
websockets==14.1
WTForms==3.1.2
